Fixed an issue with old OBs being fetched from the DB
When programs are carried over semester to semester (and maintain
the same proposal), a new set of OBs may be desired to schedule with.
But old ones left in the DB will get picked up.  This fix changes it
so that only the OBs in the files will be scheduled.